atlas_flaskapp/requirements.txt
Flask>=2.2.2
folium
networkx
requests
requests-oauthlib
google-cloud-bigquery==3.23.0
pandas
